# gateway/internal/infrastructure/api/gin/handler/unregisterSeller.go
package handler

import (
	"github.com/gin-gonic/gin"
)

type GinUnregisterSeller struct {
	//UnregisterSellerEvent *action.UnregisterSellerEvent
}

//func NewGinUnregisterSeller(unregisterSellerEvent *action.UnregisterSellerEvent) *GinUnregisterSeller {
//	return &GinUnregisterSeller{UnregisterSellerEvent: unregisterSellerEvent}
//}

// Execute Unregister a seller
// @Summary Unregister a seller
// @Description Unregister seller from a seller
// @Produce json
// @Tags Sellers
// @Param 	id 	path  string true "ID from seller to unregister"
// @Success 204
// @Failure 400
// @Failure 404
// @Failure 500
// @Failure 503
// @Router /sellers/{id} [DELETE]
func (handler GinUnregisterSeller) Execute(c *gin.Context) {
	//id, err := uuid.Parse(c.Param("id"))
	//
	//if err != nil {
	//	c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
	//	return
	//}
	//
	//err = handler.UnregisterSellerEvent.Execute(id)
	//
	//if err != nil {
	//	application.MeliGinHandlerError{}.Execute(err, c)
	//	return
	//}
	//
	//c.Status(http.StatusNoContent)
}